feat: added bytes, serde, std, taggedlen files
